requirements.txt
zetascale
swarms
PyGithub
